refs #77 Restructuring still in progress.
