--- fitbenchmarking/fitting/algorithm_fit_utils/__init__.py
@@ -1,1 +1,0 @@
-
