+++ fitbenchmarking/fitting/plotting/plotHelper.py
@@ -18,10 +18,12 @@
 # File change history is stored at: <https://github.com/mantidproject/mantid>.
 # Code Documentation is available at: <http://doxygen.mantidproject.org>
 from __future__ import (absolute_import, division, print_function)
+
 import numpy as np
 import matplotlib.pyplot as plt
 import copy
-from logging_setup import logger
+from utils.logging_setup import logger
+
 
 class data:
     """
