+++ fitbenchmarking/resprocessing/color_definitions.txt
